requirements.txt
# CT Studio Order Entry System - Dependencies

# Web Framework
streamlit>=1.28.0

# Google APIs
gspread>=5.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Optional: For better performance
protobuf>=4.24.0
